# remove deepcopy from fixedpoint