@slundqui 2023-05-31
One potential source of the slowdown is differing behavior of checkpointing between the old and new versions. Specifically, after running the [example notebook](https://github.com/delvtech/elf-simulations/blob/main/examples/hyperdrive_demo_notebook.py), the final number of elements in `simulator.market.market_state.checkpoints` went from 98 elements to 286 elements before and after https://github.com/delvtech/elf-simulations/pull/458, likely due to the calculation of `latest checkpoint time`:

https://github.com/delvtech/elf-simulations/blob/9b7c228b00a604f8abc4310d0fc7c9915d4902c7/elfpy/markets/hyperdrive/hyperdrive_market.py#L242

While this may not be the entire story, it makes sense that the speed issues come from deep copying the market state, and the checkpointing bug likely exacerbated the issue.