@dpaiton 2023-05-31
`latest_checkpoint_time` returns a FixedPoint, and so in itself should have no effect on the size of checkpoint dictionaries